requirements.txt
langchain-google-genai
python-dotenv
langchain-core
langgraph
